+++ App/xcbuddykit/Sources/Loader/Graph.swift
@@ -38,4 +38,5 @@ class Graph: Graphing {
         self.cache = cache
         self.entryNodes = entryNodes
     }
+    
 }