+++ App/xcbuddykit/Tests/Generator/ProjectGroupsTests.swift
@@ -17,7 +17,7 @@ final class ProjectGroupsTests: XCTestCase {
                               settings: nil,
                               targets: [])
         let pbxproj = PBXProj()
-        subject = ProjectGroups.generate(project: project, pbxproj: pbxproj, sourceRootPath: sourceRootPath)
+        subject = ProjectGroups.generate(project: project, objects: pbxproj.objects, sourceRootPath: sourceRootPath)
     }
 
     func test_generate() {
@@ -25,20 +25,10 @@ final class ProjectGroupsTests: XCTestCase {
         XCTAssertEqual(main.path, ".")
         XCTAssertEqual(main.sourceTree, .group)
 
-        XCTAssertTrue(main.children.contains(subject.targets.reference))
-        XCTAssertEqual(subject.targets.name, "Targets")
-        XCTAssertNil(subject.targets.path)
-        XCTAssertEqual(subject.targets.sourceTree, .group)
-
-        XCTAssertTrue(main.children.contains(subject.configurations.reference))
-        XCTAssertEqual(subject.configurations.name, "Configurations")
-        XCTAssertNil(subject.configurations.path)
-        XCTAssertEqual(subject.configurations.sourceTree, .group)
-
-        XCTAssertTrue(main.children.contains(subject.support.reference))
-        XCTAssertEqual(subject.support.name, "Support")
-        XCTAssertNil(subject.support.path)
-        XCTAssertEqual(subject.support.sourceTree, .group)
+        XCTAssertTrue(main.children.contains(subject.project.reference))
+        XCTAssertEqual(subject.project.name, "Project")
+        XCTAssertNil(subject.project.path)
+        XCTAssertEqual(subject.project.sourceTree, .group)
 
         XCTAssertTrue(main.children.contains(subject.projectDescription.reference))
         XCTAssertEqual(subject.projectDescription.name, "ProjectDescription")
@@ -56,31 +46,10 @@ final class ProjectGroupsTests: XCTestCase {
         XCTAssertEqual(subject.products.sourceTree, .buildProductsDir)
     }
 
-    func test_targetName() throws {
-        let got = try subject.target(name: "Test")
-        XCTAssertEqual(got.name, "Test")
-        XCTAssertEqual(got.sourceTree, .group)
-        XCTAssertTrue(subject.targets.children.contains(got.reference))
-    }
-
     func test_targetFrameworks() throws {
         let got = try subject.targetFrameworks(target: "Test")
         XCTAssertEqual(got.name, "Test")
         XCTAssertEqual(got.sourceTree, .group)
         XCTAssertTrue(subject.frameworks.children.contains(got.reference))
-    }
-
-    func test_targetConfigurations() throws {
-        let got = try subject.targetConfigurations("Test")
-        XCTAssertEqual(got.name, "Test")
-        XCTAssertEqual(got.sourceTree, .group)
-        XCTAssertTrue(subject.configurations.children.contains(got.reference))
-    }
-
-    func test_projectConfigurations() throws {
-        let got = try subject.projectConfigurations()
-        XCTAssertEqual(got.name, "Project")
-        XCTAssertEqual(got.sourceTree, .group)
-        XCTAssertTrue(subject.configurations.children.contains(got.reference))
     }
 }